refactor(extensions): remove HasError request extensions and move Success extensions to Response
