+++ Turner.Infrastructure.Mediator/Response.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
+using System.Threading.Tasks;
 
 namespace Turner.Infrastructure.Mediator
 {
@@ -8,6 +9,10 @@ namespace Turner.Infrastructure.Mediator
         public bool HasErrors => Errors.Any();
 
         public List<Error> Errors { get; set; } = new List<Error>();
+
+        public static Response Success() => new Response();
+
+        public static Task<Response> SuccessAsync() => Task.FromResult(Success());
     }
 
     public class Response<TResult> : Response
